Update.LICENSE.md
When using my I.P rights or my properties and making profits of them without paying or atleast asking if its okey to use my system meanwhile earning money in a lesser amount and not wanting to contribute or pay a license fee then a first warning will reach out with an email that its important to not alter or mix my codes without asking. Those who wants to contribute needs to send an message, a curriculum vitae. When it comes to them who are making profits its a must to pay the license fee